+++ src/config/settings.py
@@ -48,11 +48,19 @@ MIDDLEWARE = [
 
 ROOT_URLCONF = "config.urls"
 
+# Starting with Django 4.1+ we need to pick which template loaders to use
+# based on our environment since 4.1+ will cache templates by default.
+default_loaders = [
+    "django.template.loaders.filesystem.Loader",
+    "django.template.loaders.app_directories.Loader",
+]
+
+cached_loaders = [("django.template.loaders.cached.Loader", default_loaders)]
+
 TEMPLATES = [
     {
         "BACKEND": "django.template.backends.django.DjangoTemplates",
         "DIRS": [os.path.join(BASE_DIR, "templates")],
-        "APP_DIRS": True,
         "OPTIONS": {
             "context_processors": [
                 "django.template.context_processors.debug",
@@ -60,6 +68,7 @@ TEMPLATES = [
                 "django.contrib.auth.context_processors.auth",
                 "django.contrib.messages.context_processors.messages",
             ],
+            "loaders": default_loaders if DEBUG else cached_loaders,
         },
     },
 ]